Add option to disable copying commit message to clipboard
In the `magic_commit` script, added a new command-line argument `--no-copy` that allows the user to disable copying the generated commit message to the clipboard. This can be useful in cases where the user does not want the commit message to be automatically copied and wants to manually handle it instead.

Changes:
- Imported the `pyperclip` module to enable clipboard functionality.
- Added a new command-line argument `--no-copy` with the `action="store_true"` option to set a flag if specified.
- If the `--no-copy` flag is set, the commit message is not copied to the clipboard.
- Updated the script to use the `pyperclip.copy()` function to copy the commit message to the clipboard.
- Printed the commit message to the console regardless of whether it is copied to the clipboard or not.