Fix issue with pyperclip.copy() being called unconditionally
In the `main()` function, the `pyperclip.copy()` function was being called unconditionally, which caused an issue when the `--no-copy` command line argument was provided. To fix this, added a conditional check using the `args.no_copy` flag. If the flag is set to False, then the `pyperclip.copy()` function is called, otherwise it is skipped. This ensures that the clipboard is only copied when the `--no-copy` flag is not present.