Add contributing instructions to README
Included a new section in the README file called "Contributing" that provides instructions for users to contribute to the project. This includes information on how to open an issue or submit a pull request.

Additionally, added instructions on how to publish a new version of the project to PyPI. Users are now instructed to install `twine` if not already installed, and to run `python setup.py sdist bdist_wheel` to build the package before using `twine upload dist/*` to upload the package to PyPI.

These changes improve the project's documentation and make it easier for users to contribute and publish new versions.