Copy results to clipboard automatically from cli command (#31)
* Update setup.py with new dependencies and long description

In this commit, the `setup.py` file was updated with the following changes:
- Updated the version number to `0.4.0`
- Added `pyperclip` to the list of `install_requires` dependencies
- Removed Rahul Mysore from the `author` and `author_email` fields
- Added `long_description` field with the variable `long_description` for a more detailed project description

These changes ensure that the project is using the latest dependencies and provides a more comprehensive description of the project.

* Add option to disable copying commit message to clipboard

In the `magic_commit` script, added a new command-line argument `--no-copy` that allows the user to disable copying the generated commit message to the clipboard. This can be useful in cases where the user does not want the commit message to be automatically copied and wants to manually handle it instead.

Changes:
- Imported the `pyperclip` module to enable clipboard functionality.
- Added a new command-line argument `--no-copy` with the `action="store_true"` option to set a flag if specified.
- If the `--no-copy` flag is set, the commit message is not copied to the clipboard.
- Updated the script to use the `pyperclip.copy()` function to copy the commit message to the clipboard.
- Printed the commit message to the console regardless of whether it is copied to the clipboard or not.

* Fix issue with pyperclip.copy() being called unconditionally

In the `main()` function, the `pyperclip.copy()` function was being called unconditionally, which caused an issue when the `--no-copy` command line argument was provided. To fix this, added a conditional check using the `args.no_copy` flag. If the flag is set to False, then the `pyperclip.copy()` function is called, otherwise it is skipped. This ensures that the clipboard is only copied when the `--no-copy` flag is not present.

* Add `pyperclip` package to requirements

Added `pyperclip==1.8.2` package to the `requirements.txt` file. This package is needed for clipboard operations and is used in the project.

* Add contributing instructions to README

Included a new section in the README file called "Contributing" that provides instructions for users to contribute to the project. This includes information on how to open an issue or submit a pull request.

Additionally, added instructions on how to publish a new version of the project to PyPI. Users are now instructed to install `twine` if not already installed, and to run `python setup.py sdist bdist_wheel` to build the package before using `twine upload dist/*` to upload the package to PyPI.

These changes improve the project's documentation and make it easier for users to contribute and publish new versions.

* Run `black *.py && isort *.py` on all files