Update installation instructions in README.md
Revised the installation instructions in the README.md file to include the proper command for installing the package via PyPI. Changed the previous instruction for installing the package using Homebrew on Mac to a universal instruction for all platforms, using `pip install magic-commit`. Also added the link to the PyPI page for the package for more information.

This change improves the clarity and accuracy of the installation instructions for users.