Add alpha build of VSCode extension (#39)
* Move files into `cli` directory for monorepo setup

* Initialize the VSCode extension with `yo code`

* Rename `cli/README.md` to `README.md`

This commit renames the file `cli/README.md` to `README.md`. The content of the file remains the same, but the file location has been changed for better organization and clarity.

* Refactor `run_magic_commit()` function to add optional parameter for showing loading message

In order to make the loading message optional in the `run_magic_commit()` function, I added a new parameter called `show_loading_message` to the function's signature. This parameter is set to `True` if the `--no-load` flag is not provided.

Inside the function, I added a conditional check to determine whether to show the loading animation. If `show_loading_message` is `True`, then a new threading event is created and a separate thread is started to run the `animate_loading()` function. Otherwise, the loading animation is bypassed.

This change allows the loading message to be disabled when it is not needed, providing a cleaner and more flexible user experience.

* Add alpha build of VSCode extension

Updated the `magic-commit` extension for VSCode with the following changes:
- Changed the display name of the extension from "magic-commit" to "Magic Commit!"
- Removed unnecessary sections in the README.md file.
- Modified the `activate` function in `extension.js` to execute the `magic-commit` command and display the output in an input box as the commit message.
- Updated the version number in `package.json` from 0.0.1 to 0.0.3.
- Added the `onView:git` activation event to `package.json` to associate the extension with the Git View in VSCode.
- Updated the `contributes` section in `package.json` to include the new command for the extension and added it to the SCM (Source Control Management) title menu for easier access.

These changes improve the functionality and user experience of the `magic-commit` extension in VSCode.